+++ Code/Libraries/Foundation/Common/Math/Matrix4.cpp
@@ -84,17 +84,17 @@ Matrix4::Matrix4(ConstRealPointer data_)
   m33 = data_[15];
 }
 
-const Vector4& Matrix4::operator[](uint index) const
+const Vector4& Matrix4::operator[](size_t index) const
 {
   return ((Vector4*)this)[index];
 }
 
-Vector4& Matrix4::operator[](uint index)
+Vector4& Matrix4::operator[](size_t index)
 {
   return ((Vector4*)this)[index];
 }
 
-real Matrix4::operator()(uint r, uint c) const
+real Matrix4::operator()(size_t r, size_t c) const
 {
   ErrorIf(r > 3, "Matrix4 - Index out of range.");
   ErrorIf(c > 3, "Matrix4 - Index out of range.");
@@ -106,7 +106,7 @@ real Matrix4::operator()(uint r, uint c) const
 #endif
 }
 
-real& Matrix4::operator()(uint r, uint c)
+real& Matrix4::operator()(size_t r, size_t c)
 {
   ErrorIf(r > 3, "Matrix4 - Index out of range.");
   ErrorIf(c > 3, "Matrix4 - Index out of range.");
@@ -754,7 +754,7 @@ void Matrix4::Decompose(Mat4Param transform, Vec3Ref translation, Mat3Ref rotati
 
   // Sometimes sheer can cause the rotation matrix to not be normalized.
   // As a temp fix just force it to be normalized here.
-  for (size_t i = 0; i < 3; ++i)
+  for (uint i = 0; i < 3; ++i)
   {
     Vector3 basis = rotation.GetBasis(i);
     basis.AttemptNormalize();
